+++ docs/CHANGELOG.rst
@@ -21,6 +21,8 @@ Changed
   - ``workflow-bbduk-star-fc-quant-single``
   - ``workflow-bbduk-star-fc-quant-paired``
 - Simplify ``cellranger-count`` outputs' folder structure
+- Bump STAR aligner to version 2.7.0f in ``resolwebio/rnaseq:4.4.1``
+  Docker image
 
 Added
 -----